move locator.rs to routing module
