+++ scylla/src/connection/connection.rs
@@ -54,7 +54,7 @@ use crate::cluster::metadata::{PeerEndpoint, UntranslatedEndpoint, UntranslatedP
 use crate::cluster::NodeAddr;
 use crate::execution::errors::{ProtocolError, SchemaVersionFetchError, UseKeyspaceProtocolError};
 use crate::execution::pager::QueryPager;
-use crate::transport::locator::tablets::{RawTablet, TabletParsingError};
+use crate::routing::locator::tablets::{RawTablet, TabletParsingError};
 use crate::transport::query_result::QueryResult;
 
 use crate::batch::{Batch, BatchStatement};
