+++ scylla/src/routing/locator/replication_info.rs
@@ -204,11 +204,11 @@ where
 #[cfg(test)]
 mod tests {
     use crate::{
-        routing::Token,
-        test_utils::setup_tracing,
-        transport::locator::test::{
+        routing::locator::test::{
             create_ring, mock_metadata_for_token_aware_tests, A, B, C, D, E, F, G,
         },
+        routing::Token,
+        test_utils::setup_tracing,
     };
 
     use super::ReplicationInfo;
