+++ scylla/src/transport/mod.rs
@@ -1,5 +1,4 @@
 pub mod legacy_query_result;
-pub mod locator;
 pub mod query_result;
 
 pub use crate::frame::Authenticator;
